make automatic triggering of the search more reliable
The script previously (and erroneously) depended on the server
evaluating cookies and returning a correctly filled form (which is not
the case in some situations with a caching proxy like Varnish that just
serves the page without server contact).

Rearrange things to first evaluate cookies and then trigger the search.